[refactor] Introduce whisper_local package
