+++ speech_recognition/__init__.py
@@ -1390,13 +1390,14 @@ class PortableNamedTemporaryFile(object):
 # At this time, the dependencies are not yet installed, resulting in a ModuleNotFoundError.
 # This is a workaround to resolve this issue
 try:
-    from .recognizers import google, google_cloud, groq, openai, whisper_local
+    from .recognizers import google, google_cloud, groq, openai
+    from .recognizers.whisper_local import whisper
 except (ModuleNotFoundError, ImportError):
     pass
 else:
     Recognizer.recognize_google = google.recognize_legacy
     Recognizer.recognize_google_cloud = google_cloud.recognize
-    Recognizer.recognize_whisper = whisper_local.recognize
+    Recognizer.recognize_whisper = whisper.recognize
     Recognizer.recognize_openai = openai.recognize
     Recognizer.recognize_whisper_api = openai.recognize  # Deprecated
     Recognizer.recognize_groq = groq.recognize
